refactor: split TransactionRow, extract shared hooks, fix N+1 query
